Shift from setLong to setInt64

+++ driver/mysql_prepared_statement.h
@@ -126,7 +126,7 @@ public:
 
 	void setInt(unsigned int parameterIndex, int value);
 
-	void setLong(unsigned int parameterIndex, long long value);
+	void setInt64(unsigned int parameterIndex, int64_t value);
 
 	void setMaxFieldSize(unsigned int max);
